+++ agate/csv_py3.py
@@ -5,6 +5,7 @@ This module contains the Python 3 replacement for :mod:`csv`.
 """
 
 import csv
+import warnings
 
 import six
 
@@ -35,7 +36,7 @@ class Reader(six.Iterator):
         except csv.Error as e:
             # Terrible way to test for this exception, but there is no subclass
             if 'field larger than field limit' in str(e):
-                raise FieldSizeLimitError(csv.field_size_limit())
+                raise FieldSizeLimitError(csv.field_size_limit(), self.line_num)
             else:
                 raise e
 
@@ -151,7 +152,8 @@ class Sniffer(object):
         """
         try:
             dialect = csv.Sniffer().sniff(sample, POSSIBLE_DELIMITERS)
-        except Exception:
+        except csv.Error as e:
+            warnings.warn('Error sniffing CSV dialect: %s' % e, RuntimeWarning, stacklevel=2)
             dialect = None
 
         return dialect