+++ agate/table/from_csv.py
@@ -1,12 +1,14 @@
 #!/usr/bin/env python
 
 import io
+import itertools
 
 import six
 
 
 @classmethod
-def from_csv(cls, path, column_names=None, column_types=None, row_names=None, skip_lines=0, header=True, sniff_limit=0, encoding='utf-8', **kwargs):
+def from_csv(cls, path, column_names=None, column_types=None, row_names=None, skip_lines=0, header=True, sniff_limit=0,
+             encoding='utf-8', row_limit=None, **kwargs):
     """
     Create a new table from a CSV.
 
@@ -26,8 +28,7 @@ def from_csv(cls, path, column_names=None, column_types=None, row_names=None, sk
     :param row_names:
         See :meth:`.Table.__init__`.
     :param skip_lines:
-        The number of lines to skip from the top of the file. Note that skip
-        lines will not work with
+        The number of lines to skip from the top of the file.
     :param header:
         If :code:`True`, the first row of the CSV is assumed to contain column
         names. If :code:`header` and :code:`column_names` are both specified
@@ -39,6 +40,8 @@ def from_csv(cls, path, column_names=None, column_types=None, row_names=None, sk
         Character encoding of the CSV file. Note: if passing in a file
         handle it is assumed you have already opened it with the correct
         encoding specified.
+    :param row_limit:
+        Limit how many rows of data will be read.
     """
     from agate import csv
     from agate.table import Table
@@ -81,7 +84,10 @@ def from_csv(cls, path, column_names=None, column_types=None, row_names=None, sk
             else:
                 next(reader)
 
-        rows = tuple(reader)
+        if row_limit is None:
+            rows = tuple(reader)
+        else:
+            rows = tuple(itertools.islice(reader, row_limit))
 
     finally:
         if close: