+++ agate/table/join.py
@@ -5,13 +5,14 @@ from agate import utils
 from agate.rows import Row
 
 
-def join(self, right_table, left_key=None, right_key=None, inner=False, full_outer=False, require_match=False, columns=None):
+def join(self, right_table, left_key=None, right_key=None, inner=False, full_outer=False, require_match=False,
+         columns=None):
     """
     Create a new table by joining two table's on common values. This method
     implements most varieties of SQL join, in addition to some unique features.
 
     If :code:`left_key` and :code:`right_key` are both :code:`None` then this
-    method will peform a "sequential join", which is to say it will join on row
+    method will perform a "sequential join", which is to say it will join on row
     number. The :code:`inner` and :code:`full_outer` arguments will determine
     whether dangling left-hand and right-hand rows are included, respectively.