+++ agate/table/print_table.py
@@ -11,7 +11,8 @@ from agate import config, utils
 from agate.data_types import Number, Text
 
 
-def print_table(self, max_rows=20, max_columns=6, output=sys.stdout, max_column_width=20, locale=None, max_precision=3):
+def print_table(self, max_rows=20, max_columns=6, output=sys.stdout, max_column_width=20, locale=None,
+                max_precision=3):
     """
     Print a text-based view of the data in this table.
 
@@ -139,8 +140,6 @@ def print_table(self, max_rows=20, max_columns=6, output=sys.stdout, max_column_
 
         write('%s%s%s' % (v_line, text, v_line))
 
-    # Dashes span each width with '+' character at intersection of
-    # horizontal and vertical dividers.
     divider = '%(v_line)s %(columns)s %(v_line)s' % {
         'v_line': v_line,
         'columns': ' | '.join(h_line * w for w in widths)