+++ agate/type_tester.py
@@ -1,7 +1,9 @@
 #!/usr/bin/env python
 
+import warnings
 from copy import copy
 
+from agate.data_types.base import DEFAULT_NULL_VALUES
 from agate.data_types.boolean import Boolean
 from agate.data_types.date import Date
 from agate.data_types.date_time import DateTime
@@ -52,8 +54,11 @@ class TypeTester(object):
         options such as ``locale`` to :class:`.Number` or ``cast_nulls`` to
         :class:`.Text`. Take care in specifying the order of the list. It is
         the order they are tested in. :class:`.Text` should always be last.
+    :param null_values:
+        If :code:`types` is :code:`None`, a sequence of values which should be
+        cast to :code:`None` when encountered by the default data types.
     """
-    def __init__(self, force={}, limit=None, types=None):
+    def __init__(self, force={}, limit=None, types=None, null_values=DEFAULT_NULL_VALUES):
         self._force = force
         self._limit = limit
 
@@ -62,12 +67,12 @@ class TypeTester(object):
         else:
             # In order of preference
             self._possible_types = [
-                Boolean(),
-                Number(),
-                TimeDelta(),
-                Date(),
-                DateTime(),
-                Text()
+                Boolean(null_values=null_values),
+                Number(null_values=null_values),
+                TimeDelta(null_values=null_values),
+                Date(null_values=null_values),
+                DateTime(null_values=null_values),
+                Text(null_values=null_values)
             ]
 
     def run(self, rows, column_names):
@@ -86,7 +91,7 @@ class TypeTester(object):
             try:
                 force_indices.append(column_names.index(name))
             except ValueError:
-                raise ValueError('"%s" does not match the name of any column in this table.' % name)
+                warnings.warn('"%s" does not match the name of any column in this table.' % name, RuntimeWarning)
 
         if self._limit:
             sample_rows = rows[:self._limit]