+++ tests/test_computations.py
@@ -1,9 +1,13 @@
 #!/usr/bin/env Python
 
 import datetime
-import unittest
 import warnings
 from decimal import Decimal
+
+try:
+    import unittest2 as unittest
+except ImportError:
+    import unittest
 
 from agate import Table
 from agate.computations import Change, Formula, Percent, PercentChange, PercentileRank, Rank, Slug
@@ -50,7 +54,7 @@ class TestTableComputation(unittest.TestCase):
 
     def test_formula_invalid(self):
         with self.assertRaises(CastError):
-            new_table = self.table.compute([  # noqa
+            self.table.compute([
                 ('test', Formula(self.number_type, lambda r: r['one']))
             ])
 
@@ -175,11 +179,11 @@ class TestTableComputation(unittest.TestCase):
         self.assertEqual(to_one_place(new_table.columns['test'][3]), Decimal('60.0'))
 
         with self.assertRaises(DataTypeError):
-            new_table = self.table.compute([
+            self.table.compute([
                 ('test', Percent('two', 0))
             ])
         with self.assertRaises(DataTypeError):
-            new_table = self.table.compute([
+            self.table.compute([
                 ('test', Percent('two', -1))
             ])
         with self.assertRaises(DataTypeError):
@@ -246,12 +250,12 @@ class TestTableComputation(unittest.TestCase):
 
     def test_percent_change_invalid_columns(self):
         with self.assertRaises(DataTypeError):
-            new_table = self.table.compute([
+            self.table.compute([
                 ('test', PercentChange('one', 'three'))
             ])
 
         with self.assertRaises(DataTypeError):
-            new_table = self.table.compute([  # noqa
+            self.table.compute([
                 ('test', PercentChange('three', 'one'))
             ])