+++ tests/test_type_tester.py
@@ -1,7 +1,10 @@
 #!/usr/bin/env python
 # -*- coding: utf8 -*-
 
-import unittest
+try:
+    import unittest2 as unittest
+except ImportError:
+    import unittest
 
 from agate.data_types import Boolean, Date, DateTime, Number, Text, TimeDelta
 from agate.type_tester import TypeTester
@@ -10,6 +13,18 @@ from agate.type_tester import TypeTester
 class TestTypeTester(unittest.TestCase):
     def setUp(self):
         self.tester = TypeTester()
+
+    def test_empty(self):
+        rows = [
+            (None,),
+            (None,),
+            (None,),
+        ]
+
+        inferred = self.tester.run(rows, ['one'])
+
+        # This behavior is not necessarily desirable. See https://github.com/wireservice/agate/issues/371
+        self.assertIsInstance(inferred[0], Boolean)
 
     def test_text_type(self):
         rows = [